Restructured middleware
